bmi local server
simple bmi calculator with local hosting server. uses express and node. school assignment to grasp the concept of app.get and app.post. setting local path such as /bmicalculator which makesa  requestr and in turn app.sendfile is returned with calc.html. form data is filled out which posts back to the root which is then taken calculated and returned as simple text